# docs/ssot/api/swagger/ticketing.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fluxora Ticketing API",
    "version": "1.0.0",
    "description": "Complete backend API deployed at https://api-eticket.fluxorastudio.id (development: https://dev-api-eticket.fluxorastudio.id). Includes customer, staff, partner integration, payment callback and operational endpoints. Staff and tenant authorization are enforced per route."
  },
  "servers": [
    {
      "url": "https://dev-api-eticket.fluxorastudio.id",
      "description": "Development API (noindex deployment)"
    },
    {
      "url": "https://api-eticket.fluxorastudio.id",
      "description": "Production API hostname in current deployment configuration"
    }
  ],
...
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List public events",
        "description": "Lists published future events and public catalogue availability hints. Availability is advisory; request a quote to validate inventory.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "403": {
            "description": "Insufficient role or scope"
          },
          "404": {
            "description": "Resource unavailable or outside caller scope"
          },
          "409": {
            "description": "Conflict, expired quote, inventory or idempotency conflict"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Unexpected server error; response includes requestId"
          }
        },
        "security": []
      }
    },
    "/api/events/{slug}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get event details",
        "description": "Returns a published event by slug, including performances, ticket types and bundles.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "403": {
            "description": "Insufficient role or scope"
          },
          "404": {
            "description": "Resource unavailable or outside caller scope"
          },
          "409": {
            "description": "Conflict, expired quote, inventory or idempotency conflict"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Unexpected server error; response includes requestId"
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "Partner API"
        ],
        "summary": "List partner events",
        "description": "Requires events:read. Returns eligible published catalogue owned by the API key partner.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "403": {
            "description": "Insufficient role or scope"
          },
          "404": {
            "description": "Resource unavailable or outside caller scope"
          },
          "409": {
            "description": "Conflict, expired quote, inventory or idempotency conflict"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Unexpected server error; response includes requestId"
          }
        },
        "security": [
          {
            "partnerKey": []
          }
        ]
      }
    },
    "/api/v1/events/{slug}": {
      "get": {
        "tags": [
          "Partner API"
        ],
        "summary": "Get partner event",
        "description": "Requires events:read; cross-partner or unpublished events return 404.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "403": {
            "description": "Insufficient role or scope"
          },
          "404": {
            "description": "Resource unavailable or outside caller scope"
          },
          "409": {
            "description": "Conflict, expired quote, inventory or idempotency conflict"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Unexpected server error; response includes requestId"
          }
        },
        "security": [
          {
            "partnerKey": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "staffCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fluxora_staff"
      },
      "partnerKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Fluxora partner API key",
        "description": "Partner API key; scopes are described per operation."
      },
      "quoteToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Quote access token"
      },
      "partnerQuoteToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Order-Access-Token",
        "description": "Private quote access token for partner API; partner API key stays in Authorization."
      }
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}